+++ src/pages/PostUploadPage/PostUploadPage.jsx
@@ -2,12 +2,12 @@ import { useContext, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { handleImageSize } from "../../common/ImageResize";
 import { uploadImage } from "../../common/ImageUpload";
-import { uploadData } from "./UploadAPI";
+import { uploadData } from "./PostUploadPageAPI";
 import HeaderForm from "../../components/modules/HeaderForm/HeaderForm";
 import UploadForm from "../../components/organisms/UploadForm/UploadForm";
 import { LoginedUserContext } from "../../contexts/LoginedUserContext";
 
-function UploadPage() {
+function PostUploadPage() {
   const navigate = useNavigate();
   const [text, setText] = useState("");
   const [images, setImages] = useState([]);
@@ -47,4 +47,4 @@ function UploadPage() {
   );
 }
 
-export default UploadPage;
+export default PostUploadPage;
